Refactoring UZMultipleLayeredPopoverTouchDummyView class. Refactoring the code to dispose all popovers when tapped in the area without popovers.

+++ UZMultipleLayeredPopoverControllerSample/MyTableViewController.m
@@ -14,7 +14,7 @@
 
 @implementation MyTableViewController
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
 	return 20;
 }
